pyproject.toml
[project]
name = "eventhub-examples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.2",
    "azure-eventhub>=5.15.1",
    "azure-identity>=1.25.1",
    "azure-schemaregistry>=1.3.0",
    "azure-schemaregistry-avroencoder>=1.0.0",
    "dotenv>=0.9.9",
]
